requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.1
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ddg==0.2.2
docker==6.0.1
duckduckgo-search==2.8.5
fore==0.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
json5==0.9.11
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.27.4
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
soupsieve==2.4
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
websocket-client==1.5.1
yarl==1.8.2
youtube-transcript-api==0.5.0
fastapi~=0.95.0
uvicorn==0.21.1
playsound~=1.3.0
starlette~=0.26.1
gTTS~=2.3.1
PyObjC
pydantic~=1.10.7
selenium~=4.8.3
pinecone-client
numpy
langchain
wikipedia
chromadb